Add files via upload
ATLIQ Hardware is a company that supplies computer Hardware to multiple clients across India having Surge, Nomad, Excel, and Electricalsara Stores. As the market is growing dynamically the sales manager is facing issues in business insights. Having 3 regional managers who can provide them with the regional insights for the Products. Overall sales are declining but the manager's description doesn't give the sales manager a big Picture of what exactly is going on with the sales. They provide the Excel files with the necessary info for the sales. So we have to visualize those insights to draw a clear picture of what exactly is going on and visualize the data that can be used to get a big Picture of the sales.
This Project mainly focuses on Power BI in which I have made the dashboards. Both the Dashboards are shown in the pdf file as well given in the folder 'sales-insights-data-analysis-project'.
The folder 'sales-insights-data-analysis-project' contains the 2 SQL and 1 Power BI file.
1. 'db_dump.sql' is the file that contains the databases that need to be imported into the 'MySQL Workbench'
	and then we can employ the resultant Queries to get the desired result.
2. 'Project_1_Data_Analysis.sql' file contains the necessary and required queries used for the project to verify the results of the Dashboard made in Power BI.